functions/covid_gcs_to_bq/requirements.txt
google-cloud-bigquery==2.21.0